@@ -7,7 +7,7 @@
     "dev": "concurrently \"npm run server\" \"npm run client\"",
     "build": "npm run build-client && npm run build-server",
     "client": "webpack-dev-server --config webpack.client.dev.js",
-    "server": "nodemon src/server/index.js --exec babel-node ",
+    "server": "nodemon src/server/index.js --exec babel-node",
     "build-client": "webpack --progress --config webpack.client.prod.js",
     "build-server": "babel --no-babelrc --presets=es2015 src/server --out-dir ./dist/server",
     "serve": "NODE_ENV=production node dist/server/index.js",
@@ -20,48 +20,47 @@
   "author": "Esko Ikkala",
   "license": "MIT",
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
+    "@babel/cli": "^7.0.0",
+    "@babel/core": "^7.0.0",
+    "@babel/node": "^7.0.0",
+    "@babel/plugin-proposal-class-properties": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "@babel/preset-react": "^7.0.0",
     "babel-eslint": "^8.2.6",
     "babel-jest": "^22.4.4",
-    "babel-loader": "^7.1.5",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
+    "babel-loader": "^8.0.4",
     "clean-webpack-plugin": "^0.1.19",
     "concurrently": "^3.6.1",
     "copy-webpack-plugin": "^4.5.2",
     "css-loader": "^0.28.11",
-    "enzyme": "^3.6.0",
-    "enzyme-adapter-react-16": "^1.5.0",
+    "enzyme": "^3.7.0",
+    "enzyme-adapter-react-16": "^1.6.0",
     "eslint": "^4.19.1",
     "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jest": "^21.22.0",
-    "eslint-plugin-jsx-a11y": "^6.1.1",
+    "eslint-plugin-jest": "^21.24.1",
+    "eslint-plugin-jsx-a11y": "^6.1.2",
     "eslint-plugin-react": "^7.11.1",
     "eslint-watch": "^3.1.5",
     "file-loader": "^1.1.11",
     "html-webpack-plugin": "^3.0.7",
     "jest": "^22.4.4",
-    "jest-enzyme": "^6.0.4",
-    "react-test-renderer": "^16.5.1",
+    "jest-enzyme": "^6.1.2",
+    "react-test-renderer": "^16.5.2",
     "style-loader": "^0.20.3",
-    "webpack": "^4.19.0",
-    "webpack-cli": "^2.1.5",
-    "webpack-dev-server": "^3.1.8",
+    "webpack": "^4.20.2",
+    "webpack-cli": "^3.1.2",
+    "webpack-dev-server": "^3.1.9",
     "webpack-merge": "^4.1.4"
   },
   "dependencies": {
-    "@material-ui/core": "^3.1.0",
+    "@material-ui/core": "^3.2.0",
     "@material-ui/icons": "^3.0.1",
-    "@material-ui/lab": "^3.0.0-alpha.16",
+    "@material-ui/lab": "^3.0.0-alpha.19",
     "Leaflet.Control.Opacity": "git+https://github.com/SemanticComputing/Leaflet.Control.Opacity.git",
     "Leaflet.extra-markers": "git+https://github.com/SemanticComputing/Leaflet.ExtraMarkers.git",
     "autosuggest-highlight": "^3.1.1",
-    "babel-preset-stage-2": "^6.24.1",
     "bootstrap": "^4.1.3",
     "express": "^4.16.3",
-    "geographiclib": "^1.49.0",
     "google-maps": "^3.3.0",
     "immutable": "^3.8.2",
     "leaflet": "^1.3.4",
@@ -70,11 +69,10 @@
     "leaflet.smooth_marker_bouncing": "^1.2.0",
     "lodash": "^4.17.11",
     "node-fetch": "^2.2.0",
-    "react": "^16.5.1",
+    "react": "^16.5.2",
     "react-alert-template-basic": "^1.0.0",
-    "react-autosuggest": "^9.4.1",
     "react-csv": "^1.0.15",
-    "react-dom": "^16.5.1",
+    "react-dom": "^16.5.2",
     "react-google-maps": "^9.4.5",
     "react-redux": "^5.0.7",
     "react-redux-toastr": "^7.4.0",
@@ -87,6 +85,8 @@
     "redux-observable": "^0.18.0",
     "redux-responsive": "^4.3.8",
     "reselect": "^3.0.1",
+    "rxjs": "^6.3.3",
+    "rxjs-compat": "^6.3.3",
     "superagent": "^3.8.3",
     "victory": "^0.26.1"
   }
