@@ -32,10 +32,7 @@ module.exports = {
       {
         test: /\.js$/,
         exclude: /node_modules/,
-        loader: 'babel-loader',
-        query: {
-          presets: ['es2015', 'react', 'stage-2']
-        },
+        use: ['babel-loader']
       },
       {
         test: /\.css$/,
